pyproject.toml
[project]
name = "bot-click"
version = "0.1.0"
description = ""
authors = [
    {name = "v.kondratov"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "pytelegrambotapi (>=4.27.0,<5.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
